remove broken dependency to govvv
the govvv project is not maintained anymore and the command is not even
used in the ci as you can see in the logs of the last gh action. You can
see the following error message "line 1: govvv: command not found". The
flags were empty so it is safe to just drop it.